Refactor: endpoint accepts body of type any and validates the type in controller and sends it to correct handler in service. Moved result type handling logic to service
